gcp/requirements.txt
google-cloud-storage==1.16.1
pillow==6.0.0